refactor: rename crypto->cryptography
otherwise there is a name conflict in chalk trying to include openssl crypto headers
